Initial changes to support consumption workflow create from templates
