+++ apps/Standalone/src/designer/app/AzureLogicAppsDesigner/Utilities/Workflow.ts
@@ -1,5 +1,5 @@
 import type { ConnectionAndAppSetting, ConnectionsData, ParametersData } from '../Models/Workflow';
-import type { ConnectionReferences } from '@microsoft/logic-apps-designer';
+import { isOpenApiSchemaVersion, type ConnectionReferences } from '@microsoft/logic-apps-designer';
 
 export class WorkflowUtility {
   public static convertToCanonicalFormat(value: string): string {
@@ -87,7 +87,7 @@ export class WorkflowUtility {
 
     try {
       return JSON.parse(result);
-    } catch (error) {
+    } catch {
       throw new Error('Failure in resolving connection parameterisation');
     }
   }
@@ -159,3 +159,46 @@ export function addOrUpdateAppSettings(settings: Record<string, string>, origina
 
   return originalSettings;
 }
+
+export const getDataForConsumption = (data: any) => {
+  const properties = data?.properties as any;
+
+  const definition = removeProperties(properties?.definition, ['parameters']);
+  const connections =
+    (isOpenApiSchemaVersion(definition) ? properties?.connectionReferences : properties?.parameters?.$connections?.value) ?? {};
+
+  const workflow = { definition, connections };
+  const connectionReferences = formatConnectionReferencesForConsumption(connections);
+  const parameters: ParametersData = formatWorkflowParametersForConsumption(properties);
+
+  return { workflow, connectionReferences, parameters };
+};
+
+const removeProperties = (obj: any = {}, props: string[] = []): object => {
+  return Object.fromEntries(Object.entries(obj).filter(([key]) => !props.includes(key)));
+};
+
+const formatConnectionReferencesForConsumption = (connectionReferences: Record<string, any>): any => {
+  return Object.fromEntries(
+    Object.entries(connectionReferences).map(([key, value]) => [key, formatConnectionReferenceForConsumption(value)])
+  );
+};
+
+const formatConnectionReferenceForConsumption = (connectionReference: any): any => {
+  const connectionReferenceCopy = { ...connectionReference };
+  connectionReferenceCopy.connection = connectionReference.connection ?? { id: connectionReference.connectionId };
+  delete connectionReferenceCopy.connectionId;
+  connectionReferenceCopy.api = connectionReference.api ?? { id: connectionReference.id };
+  delete connectionReferenceCopy.id;
+  return connectionReferenceCopy;
+};
+
+const formatWorkflowParametersForConsumption = (properties: any): ParametersData => {
+  const parameters = removeProperties(properties?.definition?.parameters, ['$connections']) as ParametersData;
+  Object.entries(properties?.parameters ?? {}).forEach(([key, parameter]: [key: string, parameter: any]) => {
+    if (parameters[key]) {
+      parameters[key].value = parameter?.value;
+    }
+  });
+  return parameters;
+};
