+++ libs/logic-apps-shared/src/utils/src/lib/models/template.ts
@@ -54,6 +54,6 @@ export interface Connection {
 
 export interface TemplateContext {
   templateId: string;
-  workflowAppName: string;
+  workflowAppName?: string;
   isMultiWorkflow: boolean;
 }
